requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0
pandas==2.1.3
matplotlib==3.8.2
plotly==5.17.0
markdown2==2.4.10
reportlab==4.0.7
seaborn==0.13.0
textblob==0.17.1 
google-generativeai 